requirements.txt
pyautogui
requests
python-dotenv
pyperclip
openai
Pillow 